+++ CHANGELOG.md
@@ -13,6 +13,15 @@
   - Calculations for stratification factors now store variable names interally
     as ids allowing referenced variables to be renamed without breaking existing
     calculations
+- **Subjects Changes**
+  - The subjects index can now be filtered by subjects that have entered,
+    unentered, and missing designs by event:
+    - `EVENT:DESIGN`, ex: `baseline:demographics`
+      - Finds subjects who have entered a demographics sheet at baseline
+    - `EVENT:DESIGN:unentered`, ex: `baseline:demographics:unentered`
+      - Finds subjects who have not entered a demographics sheet at baseline
+    - `EVENT:DESIGN:missing, ex: `baseline:demographics:missing`
+      - Finds subjects who have a demographics sheet set as missing at baseline
 - **Variable Changes**
   - Increased size of calculation input for calculated variables
   - Time of day format and time duration format can now be changed when editing
@@ -33,6 +42,8 @@
 - abbr elements are no longer underlined more than once
 - Imperial height and weight variables are now evaluated correctly in
   stratification factor calculations
+- Event design tables now link correctly to filtered lists of subjects with
+  designs that are entered, unentered, and set as missing
 
 ## 0.51.1 (March 21, 2017)